requirements.txt
Django==5.1.4
django_admin_sortable2==2.2.4
django_ckeditor==6.7.2
djangorestframework==3.15.2
python-slugify==8.0.4
